moves scoped nested set into own module, applies to WorkPackage
